requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
jinja2
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic
psycopg2-binary

